Added JWT Token on employee id

+++ src/main/java/com/bl/employeepayroll/service/EmployeePayrollService.java
@@ -7,15 +7,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.bl.employeepayroll.dto.EmployeeDTO;
+import com.bl.employeepayroll.exception.RegisterException;
 import com.bl.employeepayroll.model.Employee;
 import com.bl.employeepayroll.repository.IEmployeePayrollRepo;
+import com.bl.employeepayroll.util.TokenUtil;
 
 @Service
 public class EmployeePayrollService implements IEmployeePayrollService {
-	
+
 	@Autowired
-	private IEmployeePayrollRepo employeeRepo; 
-	
+	private IEmployeePayrollRepo employeeRepo;
+
+	@Autowired
+	TokenUtil tokenutil;
+
 	@Override
 	public Employee addEmployee(EmployeeDTO employeeDTO) {
 		Employee employee = new Employee(employeeDTO);
@@ -28,28 +33,43 @@ public class EmployeePayrollService implements IEmployeePayrollService {
 	}
 
 	@Override
-	public Employee getEmployee(long id) {
-		return employeeRepo.findById(id).get();
-	}
-
-	@Override
-	public Employee updateEmployee(long id, EmployeeDTO employeeDTO) {
-		Employee emp = employeeRepo.findById(id).get();
-		
-		emp.setName(employeeDTO.getName());
-		emp.setSalary(employeeDTO.getSalary());
-		emp.setGender(employeeDTO.getGender());
-		emp.setStartDate(employeeDTO.getStartDate());
-		emp.setNote(employeeDTO.getNote());
-		emp.setProfilePic(employeeDTO.getProfilePic());
-		emp.setDepartment(employeeDTO.getDepartment());
-		emp.setUpdated_at(LocalDateTime.now());
-		return employeeRepo.save(emp);
-	}
-
-	@Override
-	public void deleteEmployee(long id) {
+	public Employee getEmployee(String token) throws RegisterException {
+		Long id = tokenutil.decodeToken(token);
 		Employee employee = employeeRepo.findById(id).get();
-		employeeRepo.delete(employee);
+		if (employee != null)
+			return employee;
+		else
+			throw new RegisterException("user not present", 400);
+	}
+
+	@Override
+	public Employee updateEmployee(String token, EmployeeDTO employeeDTO) throws RegisterException {
+		Long id = tokenutil.decodeToken(token);
+
+		Employee emp = employeeRepo.findById(id).get();
+		System.out.println(emp);
+		if (emp != null) {
+			emp.setName(employeeDTO.getName());
+			emp.setSalary(employeeDTO.getSalary());
+			emp.setGender(employeeDTO.getGender());
+			emp.setStartDate(employeeDTO.getStartDate());
+			emp.setNote(employeeDTO.getNote());
+			emp.setProfilePic(employeeDTO.getProfilePic());
+			emp.setDepartment(employeeDTO.getDepartment());
+			emp.setUpdated_at(LocalDateTime.now());
+			return employeeRepo.save(emp);
+		} else {
+			throw new RegisterException("user not present", 400);
+		}
+	}
+
+	@Override
+	public void deleteEmployee(String token) throws RegisterException {
+		Long id = tokenutil.decodeToken(token);
+		Employee employee = employeeRepo.findById(id).get();
+		if (employee != null)
+			employeeRepo.delete(employee);
+		else
+			throw new RegisterException("user not present", 400);
 	}
 }
